+++ example/events_example.dart
@@ -1,6 +1,5 @@
 // ignore_for_file: avoid_print
 /// Demonstrate events in Synthizer.
-import 'dart:ffi';
 import 'dart:io';
 
 import 'package:dart_synthizer/dart_synthizer.dart';
@@ -8,19 +7,20 @@ import 'package:dart_synthizer/dart_synthizer.dart';
 Future<void> main() async {
   final synthizer = Synthizer()..initialize();
   final ctx = synthizer.createContext(events: true);
+  ctx.getEvents().listen((event) {
+    if (event is FinishedEvent) {
+      print('Finished: ${event.source}');
+    } else if (event is LoopedEvent) {
+      print('Looped: ${event.generator}');
+    }
+  });
   final buffer = Buffer.fromFile(synthizer, File('sound.wav'));
-  final generator = ctx.createBufferGenerator()..setBuffer(buffer);
+  final generator = ctx.createBufferGenerator(buffer: buffer)..looping = true;
   final source = ctx.createDirectSource()..addGenerator(generator);
-  var looping = true;
-  ctx.events.listen(print);
-  for (var i = 0; i < 5; i++) {
-    await Future<void>.delayed(Duration(seconds: 1));
-    print(synthizer.getObjectType(generator.handle.value));
-    generator.looping = looping;
-    looping = !looping;
-  }
+  await Future<void>.delayed(Duration(seconds: 3));
   source.destroy();
   generator.destroy();
   ctx.destroy();
   synthizer.shutdown();
+  print('Done.');
 }