+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.1.0]
+
+### Added
+
+* Keep track of all created `SynthizerObject` instances.
+* Add `registerObject` and `unregisterObject` methods on the `Synthizer` class.
+* Add `getEvent` and `getEvents` methods on `Context`.`.
+
 ## [0.0.4]
 
 ### Changed